+++ packages/pluggableWidgets/chart-playground-web/src/helpers/useV2EditorController.ts
